+++ cider.el
@@ -93,7 +93,7 @@ start the server."
   (interactive "P")
   (setq cider-current-clojure-buffer (current-buffer))
   (let* ((project (when prompt-project
-                    (ido-read-directory-name "Project: ")))
+                    (read-directory-name "Project: ")))
          (project-dir (nrepl-project-directory-for
                        (or project (nrepl-current-dir)))))
     (when (nrepl-check-for-repl-buffer nil project-dir)
@@ -126,7 +126,7 @@ A default will be included consisting of `nrepl-default-host' and
   "Select an endpoint from known endpoints.
 The returned endpoint has the label removed."
   (let ((selected-endpoint (split-string
-                            (ido-completing-read
+                            (completing-read
                              "Host: " (cider-known-endpoint-candidates)))))
     (if (= 3 (length selected-endpoint))
         (cdr selected-endpoint)